requirements.txt
flask==3.0.3
flask-cors==5.0.0
gunicorn==23.0.0
joblib==1.4.2
numpy==1.26.2
pandas==2.2.3
scikit-learn==1.6.0
lightgbm==4.5.0
evidently==0.4.40   # latest stable in early 2026
faiss-cpu~=1.11.0
sentence-transformers~=4.1.0
groq~=0.29.0
prefect~=3.6.13
mysql-connector-python~=9.3.0
Werkzeug~=3.1.3
selenium~=4.33.0
beautifulsoup4~=4.13.4
webdriver-manager~=4.0.2
tqdm~=4.67.1
shap~=0.48.0
matplotlib~=3.10.3
mlflow-skinny~=3.8.1
logger~=1.4